refactor: extract RefactoringPlan signature
Relocates the RefactoringPlan DSPy signature from `dspy_modules.py` to a new, dedicated file within the `signatures` directory.
