+++ src/robofactor/dspy_modules.py
@@ -4,21 +4,11 @@ from typing import Literal
 import dspy
 
 from robofactor.signatures.code_analysis import CodeAnalysis
+from robofactor.signatures.refactoring_plan import RefactoringPlan
 
 FAILURE_SCORE = 0.0
 TRAINING_DATA_FILE = "training.json"
 logger = logging.getLogger(__name__)
-
-
-class RefactoringPlan(dspy.Signature):
-    """
-    Formulate a high-level refactoring goal and sequential action plan based on given code and analysis.
-    Return a clear summary and an ordered list of steps.
-    """
-    code_snippet: dspy.Code[Literal["python"]] = dspy.InputField(desc="Original Python code")
-    analysis: str = dspy.InputField(desc="Code analysis summary")
-    refactoring_summary: str = dspy.OutputField(description="High-level refactoring objective")
-    plan_steps: list[str] = dspy.OutputField(description="Sequential actions to achieve refactoring")
 
 
 class RefactoredCode(dspy.Signature):
